+++ Kernel/inc/drivers/io/drive.h
@@ -21,13 +21,13 @@ typedef struct drive_device {
 
     // Function pointers
     aos_bool (*init)(struct AOS_Module* m);
-    aos_bool (*read_blk)(uint64_t cidx, int port_id, uint64_t lba, uint32_t count, void* buffer);
-    aos_bool (*write_blk)(uint64_t cidx, int port_id, uint64_t lba, uint32_t count, void* buffer);
-    aos_bool (*flush)(uint64_t cidx, int port_id);
-    aos_bool (*get_block_device)(uint64_t cidx, int port_id, struct block_device* out);
+    aos_bool (*read_blk)(uint64_t cidx, uint64_t port_id, uint64_t lba, uint32_t count, void* buffer);
+    aos_bool (*write_blk)(uint64_t cidx, uint64_t port_id, uint64_t lba, uint32_t count, void* buffer);
+    aos_bool (*flush)(uint64_t cidx, uint64_t port_id);
+    aos_bool (*get_block_device)(uint64_t cidx, uint64_t port_id, struct block_device* out);
 
     struct block_device block_dev;
-    int cur_port;
+    uint64_t cur_port;
 	uint64_t controller_idx;
 
     aos_bool active;
